Separate Table Handling Code and Integrate
