+++ Word/OoxmlWord.cs
@@ -39,7 +39,6 @@ namespace OpenXML.Words
                     _mainPart.Document = new WP.Document();
                     var tmp = new OT.DefaultTemplate();
                     tmp.CreateMainDocumentPart(_mainPart);
-                    //CreateProperties(_pkgDocument);
 
                     _numberingPart = _mainPart.NumberingDefinitionsPart;
 
@@ -65,14 +64,8 @@ namespace OpenXML.Words
             {
                 try
                 {
-                    //_ms = new MemoryStream();
                     _pkgDocument = pkg;
                     _mainPart = pkg.MainDocumentPart;
-                    //_mainPart.Document = new WP.Document();
-                    //var tmp = new OT.DefaultTemplate();
-                    //tmp.CreateMainDocumentPart(_mainPart);
-                    //CreateProperties(_pkgDocument);
-
                     _numberingPart = _mainPart.NumberingDefinitionsPart;
 
                     if (_numberingPart != null)
@@ -228,7 +221,7 @@ namespace OpenXML.Words
 
                         case FF.Table ffTable:
                             {
-                                var table = CreateTable(ffTable);
+                                var table = OoxmlTable.CreateInstance(_IDs, _numberingPart).CreateTable(ffTable);
                                 _wpBody.InsertBefore(table, sectionProperties);
                                 break;
                             }
@@ -249,60 +242,6 @@ namespace OpenXML.Words
         #endregion
 
         #region Create OpenXML Table
-        internal WP.Table CreateTable(FF.Table ffTable)
-        {
-            lock (_lockObject)
-            {
-                try
-                {
-                    var rows = ffTable.Rows.Count;
-                    var cols = ffTable.Rows[0].Cells.Count;
-
-                    var wpTable = new WP.Table(
-                        new WP.TableProperties(
-                            new WP.TableStyle() { Val = ffTable.Style } // Specify the TableStyle ID you want to apply
-                        )
-                    );
-                    var tableGrid = new WP.TableGrid();
-                    for (var i = 0; i < cols; i++)
-                    {
-                        if (ffTable.Column.Width > 0)
-                            tableGrid.Append(new WP.GridColumn { Width = ffTable.Column.Width.ToString() });
-                        else
-                            tableGrid.Append(new WP.GridColumn());
-                    }
-
-                    wpTable.Append(tableGrid);
-
-                    for (var i = 0; i < rows; i++)
-                    {
-                        var wpRow = new WP.TableRow();
-
-                        for (var j = 0; j < cols; j++)
-                        {
-                            var wpCell = new WP.TableCell();
-                            var ffCell = ffTable.Rows[i].Cells[j];
-                            foreach (var ffPara in ffCell.Paragraphs)
-                            {
-                                //CreateParagraph(ffPara));
-                                wpCell.Append(OoxmlParagraph.CreateInstance(_IDs, _numberingPart).CreateParagraph(ffPara));
-                            }
-
-                            wpRow.Append(wpCell);
-                        }
-
-                        wpTable.Append(wpRow);
-                    }
-
-                    return wpTable;
-                }
-                catch (Exception ex)
-                {
-                    var errorMessage = OWD.OoxmlDocData.ConstructMessage(ex, "Create Table");
-                    throw new Openize.Words.OpenizeException(errorMessage, ex);
-                }
-            }
-        }
         #endregion
 
         #region Create OpenXML Image
@@ -1103,7 +1042,7 @@ namespace OpenXML.Words
                                     break;
                                 }
                             case WP.Table wpTable:
-                                elements.Add(LoadTable(wpTable, sequence));
+                                elements.Add(OoxmlTable.CreateInstance(_IDs, _numberingPart).LoadTable(wpTable, sequence));
                                 sequence++;
                                 break;
                             case WP.SectionProperties wpSection:
@@ -1355,63 +1294,7 @@ namespace OpenXML.Words
 
 
         #region Load OpenXML Table
-        internal FF.Table LoadTable(WP.Table wpTable, int id)
-        {
-            lock (_lockObject)
-            {
-                try
-                {
-                    var ffRows = new List<FF.Row>();
-                    foreach (var wpRow in wpTable.Elements<WP.TableRow>())
-                    {
-                        var ffRow = new FF.Row
-                        {
-                            Cells = new List<FF.Cell>()
-                        };
-                        foreach (var wpCell in wpRow.Elements<WP.TableCell>())
-                        {
-                            var ffParas = new List<FF.Paragraph>();
-                            foreach (var paragraph in wpCell.Elements<WP.Paragraph>())
-                            {
-                                ffParas.Add(OoxmlParagraph.CreateInstance(_IDs, _numberingPart).LoadParagraph(paragraph, 0));
-                            }
-
-                            var ffCell = new FF.Cell { Paragraphs = ffParas };
-                            ffRow.Cells.Add(ffCell);
-                        }
-
-                        ffRows.Add(ffRow);
-                    }
-
-                    var ffTable = new FF.Table
-                    {
-                        Rows = ffRows,
-                        ElementId = id
-                    };
-                    var tableGrid = wpTable.Elements<WP.TableGrid>().FirstOrDefault();
-                    if (tableGrid != null)
-                    {
-                        var gridColumn = tableGrid.Elements<WP.GridColumn>().FirstOrDefault();
-                        ffTable.Column.Width = Convert.ToInt32(gridColumn.Width);
-                    }
-
-                    var tableProperties = wpTable.Descendants<WP.TableProperties>().FirstOrDefault();
-                    if (tableProperties == null) return ffTable;
-                    var tableStyle = tableProperties.TableStyle;
-                    if (tableStyle != null)
-                    {
-                        ffTable.Style = tableStyle.Val;
-                    }
-
-                    return ffTable;
-                }
-                catch (Exception ex)
-                {
-                    var errorMessage = OWD.OoxmlDocData.ConstructMessage(ex, "Load Table");
-                    throw new Openize.Words.OpenizeException(errorMessage, ex);
-                }
-            }
-        }
+        
         #endregion
 
         #region Load OpenXML Section
